requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
attrs=19.3.0=pypi_0
backcall=0.1.0=pypi_0
bleach=3.1.5=pypi_0
boto=2.49.0=pypi_0
ca-certificates=2020.1.1=0
cachetools=4.1.0=pypi_0
certifi=2020.4.5.1=py36_0
chardet=3.0.4=pypi_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
google-auth=1.15.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.2.0=pypi_0
gpy=1.9.9=pypi_0
gpyopt=1.2.6=pypi_0
grpcio=1.24.3=pypi_0
idna=2.9=pypi_0
importlib-metadata=1.6.0=pypi_0
ipykernel=5.3.0=pypi_0
ipython=7.14.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.17.0=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.15.1=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.3=pypi_0
jupyter-console=6.1.0=pypi_0
jupyter-core=4.6.3=pypi_0
kiwisolver=1.2.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
markupsafe=1.1.1=pypi_0
matplotlib=3.2.1=pypi_0
mistune=0.8.4=pypi_0
mne=0.20.5=pypi_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.6=pypi_0
ncurses=6.2=he6710b0_1
nibabel=3.1.0=pypi_0
nilearn=0.6.2=pypi_0
notebook=6.0.3=pypi_0
oauthlib=3.1.0=pypi_0
openssl=1.1.1g=h7b6447c_0
opt-einsum=3.2.1=pypi_0
packaging=20.4=pypi_0
pandocfilters=1.4.2=pypi_0
paramz=0.9.5=pypi_0
parso=0.7.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pip=20.0.2=py36_3
prometheus-client=0.7.1=pypi_0
prompt-toolkit=3.0.5=pypi_0
ptyprocess=0.6.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygments=2.6.1=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.16.0=pypi_0
python=3.6.10=h7579374_2
python-dateutil=2.8.1=pypi_0
pyzmq=19.0.1=pypi_0
qtconsole=4.7.4=pypi_0
qtpy=1.9.0=pypi_0
readline=8.0=h7b6447c_0
requests=2.23.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
scikit-learn=0.23.1=pypi_0
scipy=1.4.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=46.4.0=py36_0
sklearn=0.0=pypi_0
sqlite=3.31.1=h62c20be_1
tensorboard=2.1.1=pypi_0
tensorflow-cpu=2.1.0=pypi_0
tensorflow-estimator=2.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=2.0.0=pypi_0
tk=8.6.8=hbc83047_0
tornado=6.0.4=pypi_0
traitlets=4.3.3=pypi_0
urllib3=1.25.9=pypi_0
wcwidth=0.1.9=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.34.2=py36_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
z3=0.2.0=pypi_0
zipp=3.1.0=pypi_0
zlib=1.2.11=h7b6447c_3
